Input shaping delays are now held as integers to avoid rounding error

+++ src/Platform/Platform.h
@@ -568,7 +568,7 @@ public:
 #if SUPPORT_CAN_EXPANSION
 	void HandleRemoteGpInChange(CanAddress src, uint8_t handleMajor, uint8_t handleMinor, bool state) noexcept;
 	GCodeResult UpdateRemoteStepsPerMmAndMicrostepping(AxesBitmap axesAndExtruders, const StringRef& reply) noexcept;
-	GCodeResult UpdateRemoteInputShaping(unsigned int numImpulses, const float coefficients[], const float delays[], const StringRef& reply) const noexcept;
+	GCodeResult UpdateRemoteInputShaping(unsigned int numImpulses, const float coefficients[], const uint32_t delays[], const StringRef& reply) const noexcept;
 #endif
 
 #if SUPPORT_REMOTE_COMMANDS
